fixed the slider and add new img
fixed and finish section feedback in home page
