+++ src/components/Card.js
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import "./Card.css";
-import Grp6Popup from "./Grp6Popup";
-import CopyLink from "./CopyLink";
+import Grp6Popup from "./ShareItem/Grp6Popup";
+import CopyLink from "./ShareItem/CopyLink";
 
 
 // Popup Component with Close Button
