+++ src/components/SearchBar/SearchBar.css
@@ -1,0 +1,128 @@
+.search-container {
+    position:fixed; /* التثبيت في الأعلى */
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+    padding: 5px 10px; /* استخدام padding موحد */
+    width: 100%;
+    max-width: 100%;
+    background-color:#f8f9fa;
+    z-index: 999; /* ضمان أن تكون في الأعلى بالنسبة للعناصر الأخرى */
+    top: 68px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.SelectSearch {
+    cursor: pointer;
+    width: 20%;
+    padding: 10px;
+    font-size: 15px;
+    border-radius: 5px;
+    border: 1px solid #ddd;
+    margin: 0 10px;
+}
+
+.SelectSearch:focus {
+    outline: none;
+    border-color: #d7d7d7;
+}
+
+#search-input {
+    cursor: pointer;
+    width: 100%;
+    padding: 10px;
+    font-size: 15px;
+    border-radius: 5px 5px 0 0;
+    border: 1px solid #e5e1e1;
+    margin: 0;
+    padding-right: 30px;
+}
+
+#search-input:focus {
+    outline: none;
+    box-shadow: 0 0 5px rgba(0, 0, 0, 0.2);
+}
+
+.search-icon {
+    position: absolute;
+    right: 10px;
+    top: 50%;
+    transform: translateY(-50%);
+    font-size: 18px;
+    color: #800020;
+    font-size: 25px;
+}
+
+.search-icon:hover {
+    color: #555;
+}
+
+.search-dropdown {
+    width: 100%;
+    position: absolute;
+    top: 100%;
+    left: 0;
+    background-color: #555;
+    border: 2px solid #dddddd;
+    border-top: none;
+    z-index: 998;
+    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.8);
+    max-height: 300px;
+    overflow-y: auto;
+    padding: 10px;
+    border-radius: 0 0 10px 10px;
+    text-align: right;
+}
+
+.search-result {
+    cursor: pointer;
+    display: flex;
+    flex-direction: row-reverse;
+    align-items: center;
+    padding: 8px 4px;
+}
+
+.search-result:hover {
+    background-color: rgba(255, 255, 255, 0.129);
+}
+
+.search-result:last-child {
+    border-bottom: none;
+}
+
+.result-image {
+    width: 70px;
+    height: 70px;
+    border-radius: 5px;
+    margin-left: 15px;
+}
+
+.result-details {
+    flex: 1;
+    text-align: right;
+}
+
+.result-name {
+    color: #ffffff;
+    font-size: 16px;
+    font-weight: 700;
+    margin: 0;
+    text-align: right;
+}
+
+.result-location,
+.result-rating {
+    font-size: 15px;
+    color: rgba(255, 255, 255, 0.9);
+    margin: 2px 0;
+    text-align: right;
+}
+
+.no-results {
+    color: #ffffff;
+    font-size: 16px;
+    font-weight: 500;
+    margin: 0;
+    text-align: right;
+}
