+++ src/components/ShareItem/Grp6Popup.css
@@ -1,0 +1,32 @@
+/* Container for the share popup */
+.share-popup {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-wrap: wrap;
+    position: fixed;
+    top: 25%;
+    left: 0;
+    width: 100%;
+    height: 20%;
+    z-index: 1000;
+    padding: 30px;
+}
+
+/* Style for each share icon button */
+.share-icon {
+    margin: 5px;
+    cursor: pointer;
+}
+
+/* Hover effect for icons */
+.share-icon svg:hover {
+    transform: scale(1.05);
+}
+
+/* Style for each individual icon */
+.share-icon svg {
+    width: 60px;
+    height: 60px;
+    transition: fill 0.4s ease;
+}
