--- src/components/Slideer.css
@@ -1,105 +1,0 @@
-/* General carousel styling */
-.carousel {
-    margin-bottom: 0px !important;
-    padding: 0 !important;
-}
-
-
-#carouselExampleSlidesOnly {
-    margin-bottom: 0px !important;
-    padding-bottom: 0 !important;
-}
-
-.carousel.slide {
-    position: relative;
-    width: 100%;
-    height: 470px;
-    overflow-x: hidden;
-    overflow-y: hidden;
-    display: flex;
-}
-
-
-.carousel-inner {
-    cursor: pointer;
-    flex: 1 0 100%;
-    object-fit: cover;
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    aspect-ratio: 16/6;
-}
-
-.carousel-item img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    pointer-events: none;
-}
-
-/* For screens (between 851px and 1200px) */
-@media (max-width: 1200px) {
-    .carousel-item img {
-        height: 300px;
-    }
-
-    .carousel-indicators {
-        display: none;
-    }
-
-    .carousel.slide {
-        height: 300px;
-    }
-}
-
-/* For screens (between 766px and 850px) */
-@media (min-width: 766px) and (max-width: 915px) {
-    .carousel-item img {
-        height: 300px;
-    }
-
-    .carousel-indicators {
-        display: none;
-    }
-
-    .carousel.slide {
-        height: 300px;
-    }
-}
-
-/* For screens (up to 765px) */
-@media (max-width: 765px) {
-    .carousel-item img {
-        height: 240px;
-        object-fit: cover;
-    }
-
-    .carousel-indicators {
-        display: none;
-    }
-
-    .carousel.slide {
-        height: 240px;
-    }
-}
-
-/* For very small devices */
-@media (max-width: 480px) {
-    .carousel-item img {
-        height: 240px;
-    }
-
-    .carousel-indicators {
-        display: none;
-    }
-
-    .carousel.slide {
-        height: 240px;
-    }
-}
-  /* Hide caption for screens with width 1205px or less */
-  @media (max-width: 1205px) {
-    .carousel-caption {
-      display: none;
-    }
-  }
