--- src/pages/LazyLoadSection.js
@@ -1,36 +1,0 @@
-import React, { useState, useEffect, useRef } from "react";
-
-const LazyLoadSection = ({ children }) => {
-  const [isVisible, setIsVisible] = useState(false);
-  const sectionRef = useRef();
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      ([entry]) => {
-        if (entry.isIntersecting) {
-          setIsVisible(true);
-          observer.disconnect(); 
-        }
-      },
-      { threshold: 0.9 }
-    );
-
-    if (sectionRef.current) {
-      observer.observe(sectionRef.current);
-    }
-
-    return () => {
-      if (sectionRef.current) {
-        observer.unobserve(sectionRef.current);
-      }
-    };
-  }, []);
-
-  return (
-    <div ref={sectionRef}>
-      {isVisible && children}
-    </div>
-  );
-};
-
-export default LazyLoadSection;
